KaliSocialMediaScraper/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Web Scraping & Automation
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
lxml==4.9.3
playwright==1.40.0
requests-html==0.10.0
scrapy-playwright==0.0.34

# GitHub & Social Media Scraping
PyGithub==1.59.1
tweepy==4.14.0
instaloader==4.10.1
telethon==1.32.1
discord.py==2.3.2
praw==7.7.1
python-telegram-bot==20.7
facebook-scraper==0.2.59

# OSINT Tools
# theharvester==4.4.3  # Commented out due to version issues
shodan==1.30.1
censys==2.2.8
dnspython==2.4.2
python-whois==0.8.0
phonenumbers==8.13.25
email-validator==2.1.0
ipaddress==1.0.23

# Network Analysis
networkx==3.2.1
python-louvain
python-igraph==0.10.8
# networkit==10.1  # Temporarily commented out due to Cython compilation issues

# Data Processing & Analysis
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
wordcloud==1.9.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
bokeh==3.3.2
folium==0.15.1
geopandas==0.14.1

# Database & Storage
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
elasticsearch==8.11.0
pymongo==4.6.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# Image & Media Analysis
Pillow==10.1.0
opencv-python==4.8.1.78
face-recognition==1.3.0  # Successfully installed with dlib support
exif==1.6.0
imagehash==4.3.1
# reverse-image-search==0.1.0  # Package not available in this version

# Security & Cryptography
pycryptodome==3.19.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Threat Intelligence & OSINT
pymisp
pycti
# urlscan.io==0.1.0  # Removed, use direct API calls with requests
virustotal-api==1.1.11

# Geolocation & Mapping
geopy==2.4.1
reverse-geocoder==1.5.1
timezonefinder==6.5.9
# timezone==1.2.1  # Removed, use pytz or zoneinfo

# Machine Learning & AI
tensorflow==2.15.0
torch==2.1.1
transformers==4.36.2
sentence-transformers==2.2.2
langchain==0.1.0
openai==1.3.7

# Natural Language Processing
textstat==0.7.3
language-tool-python==2.7.1
polyglot==16.7.4
langdetect==1.0.9

# Web Technologies
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.2
markdown==3.5.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
colorama==0.4.6
pyyaml==6.0.1
toml==0.10.2
configparser==6.0.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
sphinx==7.2.6

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Performance
uvloop==0.19.0
orjson==3.9.10
ujson==5.8.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Background Tasks & Message Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# HTTP Client & Scraping
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Data Processing & Analysis
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Network & Security
dnspython==2.4.2
whois==1.20240129.2
cryptography==41.0.7

# GitHub API
PyGithub==1.59.1
github3.py==4.0.1

# Social Media APIs
tweepy==4.14.0
# instagram-scraper==1.10.5  # Removed, use instaloader
youtube-dl==2021.12.17

# Domain & Network Analysis
python-whois==0.8.0
dnspython==2.4.2
# subfinder==1.0.0  # Removed, use Go-based Subfinder externally or Python alternatives
python-nmap
whois==1.20240129.2

# Data Visualization & Reporting
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
reportlab==4.0.7

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# File Processing
openpyxl==3.1.2
xlrd==2.0.1
python-docx==1.1.0
PyPDF2==3.0.1

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Machine Learning & AI
transformers==4.36.2
torch==2.1.1
tensorflow==2.15.0

# Geospatial
# geopy==2.4.0  # Removed duplicate, only keep geopy==2.4.1
folium==0.15.1

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# Configuration & Environment
pydantic-settings==2.1.0
python-decouple==3.8

# Security & Authentication
bcrypt==4.1.2
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# API Documentation
# fastapi-openapi==0.0.1  # Removed, use FastAPI's built-in OpenAPI support
# fastapi-swagger==1.0.0  # Removed, use FastAPI's built-in Swagger UI

# Task Scheduling
apscheduler==3.10.4

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Caching
redis==5.0.1
aioredis==2.0.1

# Web Scraping Extras
lxml==4.9.3
html5lib==1.1
cssselect==1.2.0

# Network Analysis
scapy==2.5.0
pyshark==0.6

# Threat Intelligence
virustotal-api==1.1.11

# Export & Reporting
jinja2==3.1.2
weasyprint==60.2

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Data Serialization
msgpack==1.0.7
orjson==3.9.10

# Monitoring & Metrics
prometheus-client==0.19.0
statsd==4.0.1

# Development Tools
ipython==8.18.1
jupyter==1.0.0 